wip: starting benchmarks utils
