+++ crates/abd-clam/benches/utils/mod.rs
@@ -1,0 +1,5 @@
+//! Utilities for benchmarks.
+
+mod compare_permuted;
+
+pub use compare_permuted::compare_permuted;
