+++ crates/abd-clam/benches/vector_search.rs
@@ -1,14 +1,9 @@
-use abd_clam::{
-    cakes::{
-        cluster::{ParSearchable, Searchable},
-        Algorithm, OffsetBall,
-    },
-    dataset::ParDataset,
-    partition::ParPartition,
-    Ball, Cluster, FlatVec, Metric,
-};
+//! Benchmark for vector search.
+
+mod utils;
+
+use abd_clam::{cakes::OffsetBall, partition::ParPartition, Ball, Cluster, FlatVec, Metric};
 use criterion::*;
-use distances::Number;
 use rand::prelude::*;
 
 const METRICS: &[(&str, fn(&Vec<f32>, &Vec<f32>) -> f32)] = &[
@@ -54,7 +49,7 @@ fn vector_search(c: &mut Criterion) {
         let mut perm_data = data.clone();
         let perm_root = OffsetBall::par_from_ball_tree(root.clone(), &mut perm_data);
 
-        compare_permuted(
+        utils::compare_permuted(
             c,
             metric_name,
             &data,
@@ -65,80 +60,6 @@ fn vector_search(c: &mut Criterion) {
             &radii,
             &ks,
         );
-    }
-}
-
-fn compare_permuted<I, U, D, Dp>(
-    c: &mut Criterion,
-    metric_name: &str,
-    data: &D,
-    root: &Ball<U>,
-    perm_data: &Dp,
-    perm_root: &OffsetBall<U>,
-    queries: &[I],
-    radii: &[U],
-    ks: &[usize],
-) where
-    I: Send + Sync,
-    U: Number,
-    D: ParDataset<I, U>,
-    Dp: ParDataset<I, U>,
-{
-    let algs = vec![
-        Algorithm::KnnRepeatedRnn(ks[0], U::ONE.double()),
-        Algorithm::KnnBreadthFirst(ks[0]),
-        Algorithm::KnnDepthFirst(ks[0]),
-    ];
-
-    let mut group = c.benchmark_group(format!("vectors-RnnClustered-{}", metric_name));
-    group
-        .sample_size(10)
-        .sampling_mode(SamplingMode::Flat)
-        .throughput(Throughput::Elements(queries.len().as_u64()));
-
-    for &radius in radii {
-        let alg = Algorithm::RnnClustered(radius);
-
-        group.bench_with_input(BenchmarkId::new("Ball", radius), &radius, |b, _| {
-            b.iter_with_large_drop(|| root.batch_search(data, queries, alg));
-        });
-        group.bench_with_input(BenchmarkId::new("OffsetBall", radius), &radius, |b, _| {
-            b.iter_with_large_drop(|| perm_root.batch_search(perm_data, queries, alg));
-        });
-
-        group.bench_with_input(BenchmarkId::new("ParBall", radius), &radius, |b, _| {
-            b.iter_with_large_drop(|| root.par_batch_search(data, queries, alg));
-        });
-        group.bench_with_input(BenchmarkId::new("ParOffsetBall", radius), &radius, |b, _| {
-            b.iter_with_large_drop(|| perm_root.par_batch_search(perm_data, queries, alg));
-        });
-    }
-    group.finish();
-
-    for alg in &algs {
-        let mut group = c.benchmark_group(format!("vectors-{}-{}", alg.variant_name(), metric_name));
-        group
-            .sample_size(10)
-            .sampling_mode(SamplingMode::Flat)
-            .throughput(Throughput::Elements(queries.len().as_u64()));
-
-        for &k in ks {
-            let alg = alg.with_params(U::ZERO, k);
-
-            group.bench_with_input(BenchmarkId::new("Ball", k), &k, |b, _| {
-                b.iter_with_large_drop(|| root.batch_search(data, queries, alg));
-            });
-            group.bench_with_input(BenchmarkId::new("OffsetBall", k), &k, |b, _| {
-                b.iter_with_large_drop(|| perm_root.batch_search(perm_data, queries, alg));
-            });
-            group.bench_with_input(BenchmarkId::new("ParBall", k), &k, |b, _| {
-                b.iter_with_large_drop(|| root.par_batch_search(data, queries, alg));
-            });
-            group.bench_with_input(BenchmarkId::new("ParOffsetBall", k), &k, |b, _| {
-                b.iter_with_large_drop(|| perm_root.par_batch_search(perm_data, queries, alg));
-            });
-        }
-        group.finish();
     }
 }
